Fix "behatch.http_call.result_pool" service not reachable.
- Extend JsonContext instead of using FeatureContext.
- Inject parameter via behat.yml.
