+++ behat.yml
@@ -7,7 +7,8 @@ default:
         - Tests\Behat\FeatureContext
         - Behat\MinkExtension\Context\MinkContext
         - Behatch\Context\RestContext
-        - Behatch\Context\JsonContext
+        - Tests\Behat\JsonContext:
+            httpCallResultPool: '@behatch.http_call.result_pool'
   extensions:
     FriendsOfBehat\SymfonyExtension:
       kernel:
